More consistent handling of internal types, minimizes unnecessary rounding.
Hue, saturation, lightness and alpha are all stored as floats, retrieved as floats.
Hue saturation and lightness are rounded off at 5 digits.
Red, green and blue are rounded and stored as ints in constructor, retrieved as ints.
Removes a number of type conversions (intval and floatval calls).
Removes a number of return types on map functions.
Updates a number of tests where we were previously off by one from SASS.